Fix 338 (#2537)
* Make sure charge is read as int8 as in spec
Will not handle files from ChemDraw >=8.0 <8.06

* Fix typo in test

+++ src/formats/chemdrawcdx.cpp
@@ -504,7 +504,7 @@ bool ChemDrawBinaryXFormat::DoFragmentImpl(CDXReader& cdxr, OBMol* pmol,
             break;
         case kCDXProp_Atom_Charge:
           if(cdxr.GetLen()==1)
-            charge = cdxr.data().get();
+            charge = (int8_t)cdxr.data().get();
           else
             READ_INT32(cdxr.data(), charge);
           break;
